requirements.txt
# Core ML frameworks
torch>=1.9.0
torchvision>=0.10.0
onnx>=1.12.0
onnxruntime-gpu>=1.12.0

# Stable Diffusion and generative models
diffusers>=0.21.0
transformers>=4.21.0
accelerate>=0.20.0
sentencepiece>=0.2.0

# System monitoring and GPU memory measurement
psutil>=5.8.0
nvidia-ml-py3>=7.352.0

# Development and testing
pytest>=6.0.0

numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.0 